refactor(ts): migrate geo components
